Rename snackbar to toast
