--- addon/templates/components/paper-snackbar.hbs
@@ -1,5 +1,0 @@
-{{#ember-wormhole to=destinationId}}
-    {{#paper-snackbar-inner swipe=swipe  swipeToClose=swipeToClose onClose=onClose right=right left=left bottom=bottom top=top capsule=capsule}}
-      {{yield (hash text="paper-snackbar-text")}}
-    {{/paper-snackbar-inner}}
-{{/ember-wormhole}}
