+++ addon/templates/components/paper-toast-inner.hbs
