--- app/components/paper-snackbar.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-paper/components/paper-snackbar';
