+++ app/components/paper-toast-text.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-paper/components/paper-toast-text';
