+++ tests/dummy/app/router.js
@@ -34,8 +34,8 @@ Router.map(function() {
     this.route('select');
     this.route('sidenav');
     this.route('slider');
-    this.route('snackbar');
     this.route('switch');
+    this.route('toast');
     this.route('toolbar');
     this.route('virtual-repeat');
   });
