+++ tests/dummy/app/templates/application.hbs
@@ -35,8 +35,8 @@
             {{#submenu-item active=(is-active "demo.select" currentRouteName) onClick=(transition-to "demo.select")}}Select{{/submenu-item}}
             {{#submenu-item active=(is-active "demo.sidenav" currentRouteName) onClick=(transition-to "demo.sidenav")}}Sidenav{{/submenu-item}}
             {{#submenu-item active=(is-active "demo.slider" currentRouteName) onClick=(transition-to "demo.slider")}}Slider{{/submenu-item}}
-            {{#submenu-item active=(is-active "demo.snackbar" currentRouteName) onClick=(transition-to "demo.snackbar")}}Snackbar{{/submenu-item}}
             {{#submenu-item active=(is-active "demo.switch" currentRouteName) onClick=(transition-to "demo.switch")}}Switch{{/submenu-item}}
+            {{#submenu-item active=(is-active "demo.toast" currentRouteName) onClick=(transition-to "demo.toast")}}Toast{{/submenu-item}}
             {{#submenu-item active=(is-active "demo.toolbar" currentRouteName) onClick=(transition-to "demo.toolbar")}}Toolbar{{/submenu-item}}
              {{#submenu-item active=(is-active "demo.virtual-repeat" currentRouteName) onClick=(transition-to "demo.virtual-repeat")}}Virtual Repeat{{/submenu-item}}
           {{/liquid-if}}
